fix and move flux input file test
